finished dev.to fetcher

+++ src/cli.rs
@@ -13,6 +13,6 @@ pub struct Cli {
     #[clap(long, help = "Dev.to api url", default_value = "https://dev.to/api/")]
     pub dev_to_url: Url,
 
-    #[clap(long, help = "Dev.to page size", default_value = "10")]
-    pub dev_to_page: u8,
+    #[clap(long, help = "Dev.to page size", default_value = "30")]
+    pub dev_to_page: u128,
 }
